+++ index.html
@@ -9,6 +9,7 @@
     <link rel="stylesheet" href="/dist/css/main.css">
     <link rel="stylesheet" href="/src/assets/fontawsome/css/all.css">
     <script type="module" src="/src/js/index.mjs"></script>
+    <script type="module" src="/src/js/posts/posts.mjs"></script>
 </head>
 <body>
     <header class="text-bg-primary">
