+++ src/js/posts/posts.mjs
@@ -1,6 +1,10 @@
 import { storageLoad } from "../storage/localstorage.mjs";
+import { getPosts } from "./get-posts.mjs";
 
 const allPostUrl = "https://nf-api.onrender.com/api/v1/social/posts";
+
+const userToken = storageLoad("accessToken");
+        console.log("USERTOKEN: " + userToken)
 
 const user = storageLoad("profile");
 const username = user.name;
@@ -8,36 +12,5 @@ console.log(username)
 
 const userOnlyPostUrl = `https://nf-api.onrender.com/api/v1/social/profiles/${username}/posts`;
 
-export async function getPosts(url) {
-
-    try {
-        //
-
-        const userToken = storageLoad("accessToken");
-        console.log("USERTOKEN: " + userToken)
-
-        if (!localStorage.getItem("accessToken")) {
-            console.log("missing accessToken");
-        } else {
-
-            const fetchOptions = {
-                method: "GET",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${userToken}`,
-                },
-            };
-            console.log(fetchOptions)
-
-            const response = await fetch(url, fetchOptions);
-            const json = await response.json();
-            console.log(json)
-        };
-    } catch (error) {
-
-        console.log(error);
-    };
-};
-
-getPosts(allPostUrl);
-getPosts(userOnlyPostUrl);
+getPosts(allPostUrl, userToken);
+getPosts(userOnlyPostUrl, userToken);
